Add code refactor
